Redactoring
* Added input_fn_generator
* Expanded `tf.estimator.RunConfig`
* Refactoring
